+++ crates/rtlola-frontend/CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.6.2] - TBD
+## [0.7.0] - TBD
+
+### Changed
+- Include origin of stream access in `accessed_by` and `accesses` of input- and output streams.
+- Include spawn and eval edges in dependency graph and annotate access edges with origin.
 
 ### Added
 - Added `is_eval_filtered` method to stream trait.