+++ crates/rtlola-hir/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.4.2] - TBD
+## [0.5.0] - TBD
 
 ### Added
 - Added the `FeatureSelector` to check for unsupported language features.
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Applying the `abs()` function to unsigned values is now rejected by the typechecker.
 - Discrete windows are now event-based and follow the pacing of the stream they aggregate.
 - Projecting out of optional tuples is now allowed and produces an optional value.
+- Include origin of stream access in `direct_accesses_with` and `direct_accessed_by_with`.
 
 ## [0.4.1] - 02.02.2023